Remove i2 php registration
