+++ src/BlockTypesController.php
@@ -179,7 +179,6 @@ final class BlockTypesController {
 
 		if ( Package::feature()->is_experimental_build() ) {
 			$block_types[] = 'SingleProduct';
-			$block_types[] = 'CheckoutI2';
 			$block_types[] = 'MiniCart';
 		}
